Add files via upload
Added files. Theory to be reviewed later.

While the merging times of bubbles to the vortex-axis have been measured, there is yet scope to learn how a bubble deforms as it approaches the vortex axis. This would be a starting point for your studies, if you choose to download these files.